mudança no stipos de váriaveis

+++ src/lib/cirulo/circulo.h
@@ -7,12 +7,12 @@
 
 typedef void* CIRCULO;/*typedef de um ponteiro opaco para 'CIRCULO'*/
 
-CIRCULO CriarCirc(int id,int x, int y, double r,const char* corb,const char* corp);
+CIRCULO CriarCirc(int id, float x, float y, double r,const char* corb,const char* corp);
 /** 
  * @brief cria um circulo com os parametros fornecidos
  * @param id é um inteiro usado para identificar a forma criada
- * @param x inteiro utilizado para definir a posição horizontal no grafico
- * @param y inteiro utilizado para definir a posição horiznotal no grafico
+ * @param x utilizado para definir a posição horizontal no grafico
+ * @param y utilizado para definir a posição horiznotal no grafico
  * @param r um numero double que irá definir o raio de criação do circulo
  * @param corb é a cor de preenchimento da borda do circulo (ex:#000000) 
  * @param corp é a cor de preenchimento da área interior do circulo (ex:#000000)
